Fixed bug in sharing histograms across processors, which suddenly became a problem. Was using MPI_DOUBLE instead of MPI_LONG.
git-svn-id: https://svn.aps.anl.gov/AOP/oag/trunk/apps/src/elegant@23036 c1b61890-85f3-ea29-e4d6-a272fe79f168

+++ frfmode.c
@@ -285,7 +285,7 @@ void track_through_frfmode(
       if (isSlave || !notSinglePart) { 
         double *buffer;
         buffer = (double*)calloc(lastBin-firstBin+1, sizeof(double));
-        MPI_Allreduce(&Ihist[firstBin], buffer, lastBin-firstBin+1, MPI_DOUBLE, MPI_SUM, workers);
+        MPI_Allreduce(&Ihist[firstBin], buffer, lastBin-firstBin+1, MPI_LONG, MPI_SUM, workers);
         memcpy(Ihist+firstBin, buffer, sizeof(double)*(lastBin-firstBin+1));
         free(buffer);
       }
